Add a missing event handler on sync
jQuery's ajax provides three handlers: error, success and complete,
the latter being triggered whatever the result is.

Since only the first two were implemented, sometimes, no toast was
displayed on the screen on the request completion. This can happen
for example when the protocol of the decoded URL is wrong (namely,
HTTP instead of HTTPS); the server might return a 302 with some HTML
that can't be parsed as JSON.